Chore: rename file to fix typo
